+++ Modules/Intent.Modules.Common.CSharp/release-notes.md
@@ -6,6 +6,7 @@
 - Added: `IsExplicitImplementationFor(string @interface)` added to `CSharpMethod` for being able to make a method explicitly implement an interface's method.
 - Added: It is now possible to specify that `CSharpStatementBlock`s and derived types are *not* `SeparatedFromPrevious` by using the the `.SeparatedFromPrevious(false)` method.
 - Added: `CSharpClass`'s `WithBaseType` and `ExtendsClass` methods now have overloads which can take an `IEnumerable<string> genericTypeParameters`.
+- Added: `InfrastructureRegisteredEvent` which is used to register Infrastructure components for Health checks (for example).
 
 ### Version 3.3.38